renamed store-services to stores-services
